requirements.txt
webrtcvad==2.0.10
soundfile==0.12.1
numpy==1.26.4
pillow==10.4.0
requests==2.32.3
twilio==9.2.3
python-dotenv==1.0.1
